Merged PR 12: fix: Fix native module cannot be null error
I reverted back to before all of the firebase stuff and rebasing happened. Then I manually readded all on the firebase code and reinstalled the packages.

I have tested on my VM of OS X Catalina, and it is running on iOS.

Not really sure what happened during the rebasing that caused the issue, but I think the packages got messed up during all of the merges from master.