feat(scoreboard): toggle between teams and people
Tapping "Teams" toggled between teams scoreboard and users scoreboard, and vice versa. People scoreboard only shows top 10. If current user is below 10th, it adds their place to the bottom of the scoreboard.

Would like to add a visual switch next to "Teams" button later to make clear that this is a button.

This required adding getUsers() to firebase.js, which gets the entire users collection.

Also added some commenting, and moved a style sheet I'd added earlier as a test but forgot about.