feat(notifications) store under uuid with uid
Notifications are now stored in firebase under a unique device identifier
rather than a random identifer and are updated when the token chnages instead
of just being left in firebase. The current user's id is also now stored along
with the push token to enable future work with notifiication
channels.